my_mom_client/requirements.txt
api==0.0.7
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
